+ json to dict and dict to json fixed, fix files and folder configs (+ utils)
